Make SeedRng and SeedRng2 accept a u32 Also, adjusted some weird code in record_mixing.c

+++ include/random.h
@@ -47,8 +47,8 @@ static inline u16 Random2(void)
 #define ISO_RANDOMIZE2(val)(1103515245 * (val) + 12345)
 
 //Sets the initial seed value of the pseudorandom number generator
-void SeedRng(u16 seed);
-void SeedRng2(u16 seed);
+void SeedRng(u32 seed);
+void SeedRng2(u32 seed);
 
 // Advances the RNG state if it is safe to do so
 void AdvanceRandom(void);
